Implemented wraparound in model2.py when computing perception vector (using sobel x and sobel y) by padding the state grid then cropping it later.
Note that we cannot use "circular_wrap" option in conv2d as we are using the functional conv2d from torch.nn.functional which uses an older conv2d before that feature was implemented.

In addition, cellIndex in compute shader has been updated to wraparound if given indices less than 0 or greater than grid x/ grid y.
NOTE: this has changed the simulation such that the edges of the cats phase are not sharp anymore.

I also added alive masking to the output.

Currently I haven't trained the cat model long enough to get a loss below 0.5 to test if training and rendering with wraparound works.

The image is still oriented the wrong way

OTHER DETAILS:
- Working model weights without wraparound in training is called "model_weights.pth" and imported from "model_weights.bin" IMPORTANT: ensure to run save_weights.py if update model_weights.pth