Added Turbulation Sigmoid Function and have made small changes to lradj and train_lra so that the code functions without any issues.
Moved from torch module for statistics (e.g std, median) to statistics module due to issues with tensors.

To update lr of the optimiser we cannot directly update it as it has no "lr" attribute. Rather the paramter groups of the optimiser need to be updated.

Also, do not perform learning rate adjustment in first epoch.